fix: cleanup dependencies, code and deprecated cli commands (extend-docs, search, etc.)
