+++ packages-node/providence-analytics/src/program/analyzers/find-classes.js
@@ -1,10 +1,10 @@
 /* eslint-disable no-shadow, no-param-reassign */
 import path from 'path';
+
 import t from '@babel/types';
-// @ts-ignore
 import babelTraverse from '@babel/traverse';
 import { Analyzer } from '../core/Analyzer.js';
-import { trackDownIdentifierFromScope } from './helpers/track-down-identifier--legacy.js';
+import { trackDownIdentifierFromScope } from '../utils/track-down-identifier--legacy.js';
 
 /**
  * @typedef {import('@babel/types').File} File
