+++ packages-node/providence-analytics/src/program/providence.js
@@ -148,28 +148,6 @@ async function handleAnalyzer(queryConfig, cfg) {
   return queryResults;
 }
 
-async function handleFeature(queryConfig, cfg, inputData) {
-  if (cfg.queryMethod === 'grep') {
-    const queryResult = await QueryService.grepSearch(inputData, queryConfig, {
-      gatherFilesConfig: cfg.gatherFilesConfig,
-      gatherFilesConfigReference: cfg.gatherFilesConfigReference,
-    });
-    return queryResult;
-  }
-  return undefined;
-}
-
-async function handleRegexSearch(queryConfig, cfg, inputData) {
-  if (cfg.queryMethod === 'grep') {
-    const queryResult = await QueryService.grepSearch(inputData, queryConfig, {
-      gatherFilesConfig: cfg.gatherFilesConfig,
-      gatherFilesConfigReference: cfg.gatherFilesConfigReference,
-    });
-    return queryResult;
-  }
-  return undefined;
-}
-
 /**
  * Creates a report with usage metrics, based on a queryConfig.
  *
@@ -220,23 +198,7 @@ export async function providence(queryConfig, customConfig) {
     AstService.fallbackToBabel = true;
   }
 
-  let queryResults;
-  if (queryConfig.type === 'ast-analyzer') {
-    queryResults = await handleAnalyzer(queryConfig, cfg);
-  } else {
-    const inputData = await InputDataService.createDataObject(
-      cfg.targetProjectPaths,
-      cfg.gatherFilesConfig,
-    );
-
-    if (queryConfig.type === 'feature') {
-      queryResults = await handleFeature(queryConfig, cfg, inputData);
-      report(queryResults, cfg);
-    } else if (queryConfig.type === 'search') {
-      queryResults = await handleRegexSearch(queryConfig, cfg, inputData);
-      report(queryResults, cfg);
-    }
-  }
+  const queryResults = await handleAnalyzer(queryConfig, cfg);
 
   if (cfg.writeLogFile) {
     LogService.writeLogFile();
