+++ packages-node/providence-analytics/src/program/utils/from-import-to-export-perspective.js
@@ -1,13 +1,14 @@
 import path from 'path';
-import { isRelativeSourcePath } from '../../utils/relative-source-path.js';
-import { LogService } from '../../core/LogService.js';
-import { resolveImportPath } from '../../utils/resolve-import-path.js';
-import { toPosixPath } from '../../utils/to-posix-path.js';
+
+import { isRelativeSourcePath } from './relative-source-path.js';
+import { resolveImportPath } from './resolve-import-path.js';
+import { LogService } from '../core/LogService.js';
+import { toPosixPath } from './to-posix-path.js';
 
 /**
- * @typedef {import('../../../../types/index.js').PathRelativeFromProjectRoot} PathRelativeFromProjectRoot
- * @typedef {import('../../../../types/index.js').PathFromSystemRoot} PathFromSystemRoot
- * @typedef {import('../../../../types/index.js').SpecifierSource} SpecifierSource
+ * @typedef {import('../../../types/index.js').PathRelativeFromProjectRoot} PathRelativeFromProjectRoot
+ * @typedef {import('../../../types/index.js').PathFromSystemRoot} PathFromSystemRoot
+ * @typedef {import('../../../types/index.js').SpecifierSource} SpecifierSource
  */
 
 /**
