+++ packages-node/providence-analytics/src/program/utils/index.js
@@ -1,9 +1,13 @@
 export { toRelativeSourcePath, isRelativeSourcePath } from './relative-source-path.js';
+export { trackDownIdentifier } from './track-down-identifier.js';
 export {
   getSourceCodeFragmentOfDeclaration,
   getFilePathOrExternalSource,
 } from './get-source-code-fragment-of-declaration.js';
 export { optimisedGlob } from './optimised-glob.js';
+export { swcTraverse } from './swc-traverse.js';
 export { fsAdapter } from './fs-adapter.js';
 export { memoize } from './memoize.js';
+export { hash } from './hash.js';
+
 // TODO: move trackdownIdentifier to utils as well
