+++ packages-node/providence-analytics/src/program/utils/providence-conf-util.js
@@ -1,5 +1,6 @@
-import path from 'path';
 import { pathToFileURL } from 'url';
+import path from 'path';
+
 import { fsAdapter } from './fs-adapter.js';
 
 /**
