+++ packages-node/providence-analytics/src/program/utils/track-down-identifier.js
@@ -1,21 +1,19 @@
-/* eslint-disable no-shadow */
-// @ts-nocheck
-import pathLib from 'path';
-import babelTraverse from '@babel/traverse';
-import { isRelativeSourcePath, toRelativeSourcePath } from '../../utils/relative-source-path.js';
-import { InputDataService } from '../../core/InputDataService.js';
-import { resolveImportPath } from '../../utils/resolve-import-path.js';
-import { AstService } from '../../core/AstService.js';
-import { LogService } from '../../core/LogService.js';
-import { memoize } from '../../utils/memoize.js';
-import { fsAdapter } from '../../utils/fs-adapter.js';
+import path from 'path';
+
+import { isRelativeSourcePath, toRelativeSourcePath } from './relative-source-path.js';
+import { InputDataService } from '../core/InputDataService.js';
+import { resolveImportPath } from './resolve-import-path.js';
+import { AstService } from '../core/AstService.js';
+import { swcTraverse } from './swc-traverse.js';
+import { fsAdapter } from './fs-adapter.js';
+import { memoize } from './memoize.js';
 
 /**
- * @typedef {import('../../../../types/index.js').RootFile} RootFile
- * @typedef {import('../../../../types/index.js').SpecifierSource} SpecifierSource
- * @typedef {import('../../../../types/index.js').IdentifierName} IdentifierName
- * @typedef {import('../../../../types/index.js').PathFromSystemRoot} PathFromSystemRoot
- * @typedef {import('@babel/traverse').NodePath} NodePath
+ * @typedef {import('../../../types/index.js').PathFromSystemRoot} PathFromSystemRoot
+ * @typedef {import('../../../types/index.js').SpecifierSource} SpecifierSource
+ * @typedef {import('../../../types/index.js').IdentifierName} IdentifierName
+ * @typedef {import('../../../types/index.js').RootFile} RootFile
+ * @typedef {import('../../../types/index.js').SwcPath} SwcPath
  */
 
 /**
@@ -42,25 +40,28 @@ function isExternalProject(source, projectName) {
  * Other than with import, no binding is created for MyClass by Babel(?)
  * This means 'path.scope.getBinding('MyClass')' returns undefined
  * and we have to find a different way to retrieve this value.
- * @param {NodePath} astPath Babel ast traversal path
+ * @param {SwcPath} swcPath Babel ast traversal path
  * @param {IdentifierName} identifierName the name that should be tracked (and that exists inside scope of astPath)
  */
-function getBindingAndSourceReexports(astPath, identifierName) {
+function getBindingAndSourceReexports(swcPath, identifierName) {
   // Get to root node of file and look for exports like `export { identifierName } from 'src';`
   let source;
   let bindingType;
   let bindingPath;
 
-  let curPath = astPath;
+  let curPath = swcPath;
   while (curPath.parentPath) {
     curPath = curPath.parentPath;
   }
   const rootPath = curPath;
-  rootPath.traverse({
+
+  swcTraverse(rootPath.node, {
     ExportSpecifier(astPath) {
       // eslint-disable-next-line arrow-body-style
       const found =
-        astPath.node.exported.name === identifierName || astPath.node.local.name === identifierName;
+        astPath.node.orig?.value === identifierName ||
+        astPath.node.exported?.value === identifierName ||
+        astPath.node.local?.value === identifierName;
       if (found) {
         bindingPath = astPath;
         bindingType = 'ExportSpecifier';
@@ -80,7 +81,7 @@ function getBindingAndSourceReexports(astPath, identifierName) {
  * We might be an import that was locally renamed.
  * Since we are traversing, we are interested in the imported name. Or in case of a re-export,
  * the local name.
- * @param {NodePath} astPath Babel ast traversal path
+ * @param {SwcPath} astPath Babel ast traversal path
  * @param {string} identifierName the name that should be tracked (and that exists inside scope of astPath)
  * @returns {{ source:string, importedIdentifierName:string }}
  */
@@ -88,7 +89,7 @@ export function getImportSourceFromAst(astPath, identifierName) {
   let source;
   let importedIdentifierName;
 
-  const binding = astPath.scope.getBinding(identifierName);
+  const binding = astPath.scope.bindings[identifierName];
   let bindingType = binding?.path.type;
   let bindingPath = binding?.path;
   const matchingTypes = ['ImportSpecifier', 'ImportDefaultSpecifier', 'ExportSpecifier'];
@@ -105,7 +106,7 @@ export function getImportSourceFromAst(astPath, identifierName) {
     importedIdentifierName = '[default]';
   } else if (source) {
     const { node } = bindingPath;
-    importedIdentifierName = (node.imported && node.imported.name) || node.local.name;
+    importedIdentifierName = node.orig?.value || node.imported?.value || node.local?.value;
   }
 
   return { source, importedIdentifierName };
@@ -175,17 +176,26 @@ async function trackDownIdentifierFn(
 
   const resolvedSourcePath = await resolveImportPath(source, currentFilePath);
 
-  LogService.debug(`[trackDownIdentifier] ${resolvedSourcePath}`);
+  // if (resolvedSourcePath === null) {
+  //   LogService.error(`[trackDownIdentifier] ${resolvedSourcePath} not found`);
+
+  // }
+  // if (resolvedSourcePath === '[node-builtin]') {
+  //   LogService.error(`[trackDownIdentifier] ${resolvedSourcePath} not found`);
+  // }
+
   const allowedJsModuleExtensions = ['.mjs', '.js'];
-  if (!allowedJsModuleExtensions.includes(pathLib.extname(resolvedSourcePath))) {
+  if (
+    !allowedJsModuleExtensions.includes(path.extname(/** @type {string} */ (resolvedSourcePath)))
+  ) {
     // We have an import assertion
     return /** @type { RootFile } */ {
-      file: toRelativeSourcePath(resolvedSourcePath, rootPath),
+      file: toRelativeSourcePath(/** @type {string} */ (resolvedSourcePath), rootPath),
       specifier: '[default]',
     };
   }
-  const code = fsAdapter.fs.readFileSync(resolvedSourcePath, 'utf8');
-  const babelAst = AstService.getAst(code, 'swc-to-babel', { filePath: resolvedSourcePath });
+  const code = fsAdapter.fs.readFileSync(/** @type {string} */ (resolvedSourcePath), 'utf8');
+  const swcAst = AstService._getSwcAst(code);
 
   const shouldLookForDefaultExport = identifierName === '[default]';
 
@@ -193,96 +203,111 @@ async function trackDownIdentifierFn(
   let exportMatch;
   let pendingTrackDownPromise;
 
-  babelTraverse.default(babelAst, {
-    ExportDefaultDeclaration(astPath) {
-      if (!shouldLookForDefaultExport) {
+  const handleExportDefaultDeclOrExpr = astPath => {
+    if (!shouldLookForDefaultExport) {
+      return;
+    }
+
+    let newSource;
+    if (
+      astPath.node.expression?.type === 'Identifier' ||
+      astPath.node.declaration?.type === 'Identifier'
+    ) {
+      newSource = getImportSourceFromAst(astPath, astPath.node.expression.value).source;
+    }
+
+    if (newSource) {
+      pendingTrackDownPromise = trackDownIdentifier(
+        newSource,
+        '[default]',
+        /** @type {PathFromSystemRoot} */ (resolvedSourcePath),
+        rootPath,
+        projectName,
+        depth + 1,
+      );
+    } else {
+      // We found our file!
+      rootSpecifier = identifierName;
+      rootFilePath = toRelativeSourcePath(
+        /** @type {PathFromSystemRoot} */ (resolvedSourcePath),
+        rootPath,
+      );
+    }
+    astPath.stop();
+  };
+  const handleExportDeclOrNamedDecl = {
+    enter(astPath) {
+      if (reexportMatch || shouldLookForDefaultExport) {
         return;
       }
 
-      let newSource;
-      if (astPath.node.declaration.type === 'Identifier') {
-        newSource = getImportSourceFromAst(astPath, astPath.node.declaration.name).source;
-      }
-
-      if (newSource) {
-        pendingTrackDownPromise = trackDownIdentifier(
-          newSource,
-          '[default]',
-          resolvedSourcePath,
-          rootPath,
-          projectName,
-          depth + 1,
+      // Are we dealing with a re-export ?
+      if (astPath.node.specifiers?.length) {
+        exportMatch = astPath.node.specifiers.find(
+          s => s.orig?.value === identifierName || s.exported?.value === identifierName,
         );
-      } else {
-        // We found our file!
-        rootSpecifier = identifierName;
-        rootFilePath = toRelativeSourcePath(resolvedSourcePath, rootPath);
-      }
-      astPath.stop();
-    },
-    ExportNamedDeclaration: {
-      enter(astPath) {
-        if (reexportMatch || shouldLookForDefaultExport) {
-          return;
-        }
 
-        // Are we dealing with a re-export ?
-        if (astPath.node.specifiers?.length) {
-          exportMatch = astPath.node.specifiers.find(s => s.exported.name === identifierName);
+        if (exportMatch) {
+          const localName = exportMatch.orig.value;
+          let newSource;
+          if (astPath.node.source) {
+            /**
+             * @example
+             * export { x } from 'y'
+             */
+            newSource = astPath.node.source.value;
+          } else {
+            /**
+             * @example
+             * import { x } from 'y'
+             * export { x }
+             */
+            newSource = getImportSourceFromAst(astPath, identifierName).source;
 
-          if (exportMatch) {
-            const localName = exportMatch.local.name;
-            let newSource;
-            if (astPath.node.source) {
+            if (!newSource || newSource === '[current]') {
               /**
                * @example
-               * export { x } from 'y'
-               */
-              newSource = astPath.node.source.value;
-            } else {
-              /**
-               * @example
-               * import { x } from 'y'
+               * const x = 12;
                * export { x }
                */
-              newSource = getImportSourceFromAst(astPath, identifierName).source;
-
-              if (!newSource || newSource === '[current]') {
-                /**
-                 * @example
-                 * const x = 12;
-                 * export { x }
-                 */
-                return;
-              }
+              return;
             }
-            reexportMatch = true;
-            pendingTrackDownPromise = trackDownIdentifier(
-              newSource,
-              localName,
-              resolvedSourcePath,
-              rootPath,
-              projectName,
-              depth + 1,
-            );
-            astPath.stop();
           }
+          reexportMatch = true;
+          pendingTrackDownPromise = trackDownIdentifier(
+            newSource,
+            localName,
+            resolvedSourcePath,
+            rootPath,
+            projectName,
+            depth + 1,
+          );
+          astPath.stop();
         }
-      },
-      exit(astPath) {
-        if (!reexportMatch) {
-          // We didn't find a re-exported Identifier, that means the reference is declared
-          // in current file...
-          rootSpecifier = identifierName;
-          rootFilePath = toRelativeSourcePath(resolvedSourcePath, rootPath);
-
-          if (exportMatch) {
-            astPath.stop();
-          }
-        }
-      },
+      }
     },
-  });
+    exit(astPath) {
+      if (!reexportMatch) {
+        // We didn't find a re-exported Identifier, that means the reference is declared
+        // in current file...
+        rootSpecifier = identifierName;
+        rootFilePath = toRelativeSourcePath(resolvedSourcePath, rootPath);
+
+        if (exportMatch) {
+          astPath.stop();
+        }
+      }
+    },
+  };
+
+  const visitor = {
+    ExportDefaultDeclaration: handleExportDefaultDeclOrExpr,
+    ExportDefaultExpression: handleExportDefaultDeclOrExpr,
+    ExportNamedDeclaration: handleExportDeclOrNamedDecl,
+    ExportDeclaration: handleExportDeclOrNamedDecl,
+  };
+
+  swcTraverse(swcAst, visitor, { needsAdvancedPaths: true });
 
   if (pendingTrackDownPromise) {
     // We can't handle promises inside Babel traverse, so we do it here...
@@ -297,7 +322,7 @@ async function trackDownIdentifierFn(
 trackDownIdentifier = memoize(trackDownIdentifierFn);
 
 /**
- * @param {NodePath} astPath
+ * @param {SwcPath} astPath
  * @param {string} identifierNameInScope
  * @param {PathFromSystemRoot} fullCurrentFilePath
  * @param {PathFromSystemRoot} projectPath
